Fix #302

+++ wstx/branches/4.4/src/java/com/ctc/wstx/io/BaseInputSource.java
@@ -166,6 +166,6 @@ public abstract class BaseInputSource
          *   let's just let that be...
          */
         return new WstxInputLocation(pl, getPublicId(), getSystemId(),
-                                     (int) total, row, col);
+                                     total, row, col);
     }
 }
